requirements.txt
# AI Daily News — 四层流水线依赖
# Python >= 3.10
# 推荐使用: /Users/niu/.workbuddy/binaries/python/versions/3.14.3/bin/python3

# ── Layer 1: 收集（collector） ──
feedparser>=6.0.11         # RSS/Atom 解析
exa-py>=1.1.0              # Exa 搜索 API（无 RSS 网站兜底）
# Twitter/X 采集通过 xreach CLI（agent-reach skill），无需 Python 包
httpx>=0.27.0              # 异步 HTTP 客户端（带超时/重试）

# ── Layer 2: 筛选（filter） ──
beautifulsoup4>=4.12.3     # HTML 清洗
python-dateutil>=2.9.0     # 日期解析

# ── Layer 3: 编辑（editor） ──
# LLM 摘要/分类由 CodeBuddy 在对话中完成，不依赖外部 API
# jinja2>=3.1              # 模板渲染（可选）

# ── Layer 4: 归档（archiver） ──
# 无额外依赖

# ── 通用 ──
pyyaml>=6.0.1              # YAML 配置解析
python-dotenv>=1.0.1       # 环境变量（.env 文件）
pydantic>=2.7.0            # 数据校验
rich>=13.7.0               # 美化终端输出
